pyproject.toml
[project]
name = "coifnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "mlflow>=2.22.1",
    "psutil>=7.0.0",
    "thop>=0.1.1.post2209072238",
    "torch>=2.7.0",
    "torchmetrics>=1.7.2",
    "tqdm>=4.67.1",
]
